+++ NEWS.md
@@ -3,6 +3,8 @@
 * Added back code generation in the app (contributed by Dean Attali)
 * bumped the version to 1.0 to correspond with the tutorial paper
 * added back pk and pkpd vignettes after fixing R 4.0 issues
+* added break_y_label and y_label_text_width arguments to break long y axis labels
+* fixed a bug with show_table_yaxis_tick_label
 
 # coveffectsplot 0.0.9